formatting/requirements.txt
streamlit>=1.7.0
python-docx==0.8.11
pymupdf>=1.23.0
pdf2image>=1.16.0
Pillow>=9.0.0
pytesseract>=0.3.10
altair>=4.0,<5
click>=8.0,<8.1
openai>=1.0.0
python-dotenv>=1.0.0
streamlit-quill>=0.0.3
streamlit-lexical>=1.0.0
markdown>=3.4.0
mammoth>=1.6.0
beautifulsoup4>=4.12.0